IRI.pm:  Must specify lower bound on regex quantifiers
There are 5 occurrences of /{,\d}/ in this module.  I suspect these were
meant to mean {0,\d}, that is a quantifier matching 0 times up to the
specified limit.  But instead Perl thinks they are trying to match a
left brace followed by a comma, followed by the digits, followed by a
right brace.  Perl 5.25 now warns about this usage, and so this was
found.